requirements.txt
Pillow==9.4.0
pyperclip==1.7.0
